Add styling to notes

+++ apps/note/pages/NoteIndex.jsx
@@ -1,22 +1,28 @@
 
-// const { Link, useSearchParams } = ReactRouterDOM
+const { Link, useSearchParams } = ReactRouterDOM
 
+import { NoteFilter } from "../cmps/NoteFilter.jsx"
 import { NoteList } from "../cmps/NoteList.jsx"
 import { noteService } from "../services/note.service.js"
-// import { showErrorMsg, showSuccessMsg } from "../services/event-bus.service.js"
+// import { showErrorMsg, showSuccessMsg } from "../../../services/event-bus.service.js"
 
 const { useEffect, useState } = React
+const { Outlet } = ReactRouterDOM
 
 export function NoteIndex() {
+
+    const [searchParams, setSearchParams] = useSearchParams()
     const [notes, setNotes] = useState(null)
+    const [filterBy, setFilterBy] = useState(noteService.getFilterFromSearchParams(searchParams))
 
 
     useEffect(() => {
         loadNotes()
-    }, [])
+        setSearchParams(filterBy)
+    }, [filterBy])
 
     function loadNotes() {
-        noteService.query()
+        noteService.query(filterBy)
             .then(notes => {
                 setNotes(notes)
             })
@@ -26,7 +32,8 @@ export function NoteIndex() {
     }
 
     function onRemoveNote(noteId) {
-        noteService.remove(noteId)
+        if(confirm('are you sure?')){
+            noteService.remove(noteId)
             .then(() => {
                 setNotes(notes =>
                     notes.filter(note => note.id !== noteId)
@@ -35,22 +42,26 @@ export function NoteIndex() {
             })
             .catch(err => {
                 console.log('Problems removing note:', err)
-                showErrorMsg(`Having problems removing car!`)
+                showErrorMsg(`Having problems removing note!`)
             })
+        }
     }
 
-    // function onSetFilter(filterBy) {
-    //     setFilterBy(prevFilter => ({ ...prevFilter, ...filterBy }))
-    //     // setFilterBy({ ...filterBy })
-    // }
+    function onSetFilter(filterBy) {
+        setFilterBy(prevFilter => ({ ...prevFilter, ...filterBy }))
+        // setFilterBy({ ...filterBy })
+    }
 
     if (!notes) return <div>Loading...</div>
 
     return (
         <section className="note-index">
-            <NoteList 
-            notes={notes}
-            onRemoveNote={onRemoveNote} />
+            {/* <h1> search</h1> */}
+            <NoteFilter filterBy={filterBy} onSetFilter={onSetFilter} />
+            <NoteList
+                notes={notes}
+                onRemoveNote={onRemoveNote} />
+                <Outlet/>
         </section>
     )
 }
